TASK: Move `GlobalRegistry` "validators" overload to validators package
... and also instantiate validators registry here
