+++ packages/neos-ui-redux-store/src/UI/Inspector/selectors.ts
@@ -1,9 +1,10 @@
 import {createSelector, defaultMemoize} from 'reselect';
 
 import validate from '@neos-project/neos-ui-validators';
+import type {ValidatorRegistry} from '@neos-project/neos-ui-validators';
 import {selectors as nodes} from '../../CR/Nodes';
 import {GlobalState} from '../../..';
-import {NodeTypesRegistry, ValidatorRegistry} from '@neos-project/neos-ts-interfaces';
+import {NodeTypesRegistry} from '@neos-project/neos-ts-interfaces';
 
 /**
  * Transient values and everything which depends on this
